requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
chardet==4.0.0
charset-normalizer==2.1.0
discord-py-interactions==4.3.1
discord.py==1.7.3
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
typing_extensions==4.3.0
yarl==1.7.2